crawlers/requirements.txt
scrapy==2.11.0
newspaper3k==0.2.8
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
python-dateutil==2.8.2
pytz==2023.3
mysql-connector-python==8.1.0
python-dotenv==1.0.0
argparse==1.4.0
json-logging==1.3.0